+++ test/intern.base-config.js
@@ -92,6 +92,7 @@ define([
       'test/unit/is.visible.test',
       'test/unit/maintain.disabled.test',
       'test/unit/maintain.focus-trapped.test',
+      'test/unit/maintain.hidden.test',
       'test/unit/observe.interaction-type.test',
       'test/unit/prototype.element.prototype.matches.test',
       'test/unit/prototype.svgelement.prototype.focus.test',