+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * adding `element/disabled` (refactored from `focus/disabled`) - [issue #33](https://github.com/medialize/ally.js/issues/33)
 * adding `when/key` to observe simple keyboard input - [issue #47](https://github.com/medialize/ally.js/issues/47)
 * adding `get/insignificant-branches` to find branches not relevant to a given set of elements - [issue #32](https://github.com/medialize/ally.js/issues/32)
+* adding `maintain/hidden` to hide everything in the DOM that is not relevant to a given set of elements - [issue #46](https://github.com/medialize/ally.js/issues/46)
 
 ### Breaking Changes